refactor argument parser in separate function
