+++ tutorials/islands_example.py
@@ -1,6 +1,5 @@
-#!/usr/bin/env python3
-import argparse
-import logging
+"""Simple island model example script."""
+import pathlib
 import random
 
 import numpy as np
@@ -9,75 +8,18 @@ from mpi4py import MPI
 from propulate import Islands
 from propulate.propagators import SelectMin, SelectMax
 from propulate.utils import get_default_propagator, set_logger_config
-from function_benchmark import get_function_search_space
+from function_benchmark import get_function_search_space, parse_arguments
 
 if __name__ == "__main__":
     comm = MPI.COMM_WORLD
 
-    parser = argparse.ArgumentParser(
-        prog="Simple Propulator example",
-        description="Set up and run a basic Propulator optimization of mathematical functions.",
-    )
-    parser.add_argument(  # Function to optimize
-        "-f",
-        "--function",
-        type=str,
-        choices=[
-            "bukin",
-            "eggcrate",
-            "himmelblau",
-            "keane",
-            "leon",
-            "rastrigin",
-            "schwefel",
-            "sphere",
-            "step",
-            "rosenbrock",
-            "quartic",
-            "bisphere",
-            "birastrigin",
-            "griewank",
-        ],
-        default="sphere",
-    )
-    parser.add_argument(
-        "-g", "--generations", type=int, default=1000
-    )  # Number of generations
-    parser.add_argument(
-        "-s", "--seed", type=int, default=0
-    )  # Seed for Propulate random number generator
-    parser.add_argument("-v", "--verbosity", type=int, default=1)  # Verbosity level
-    parser.add_argument(
-        "-ckpt", "--checkpoint", type=str, default="./"
-    )  # Path for loading and writing checkpoints.
-    parser.add_argument(
-        "-p", "--pop_size", type=int, default=2 * comm.size
-    )  # Breeding pool size
-    parser.add_argument(
-        "-cp", "--crossover_probability", type=float, default=0.7
-    )  # Crossover probability
-    parser.add_argument(
-        "-mp", "--mutation_probability", type=float, default=0.4
-    )  # Mutation probability
-    parser.add_argument("-rp", "--random_init_probability", type=float, default=0.1)
-    parser.add_argument(
-        "-i", "--num_islands", type=int, default=2
-    )  # Number of separate evolutionary islands
-    parser.add_argument(
-        "-migp", "--migration_probability", type=float, default=0.9
-    )  # Migration probability
-    parser.add_argument("-m", "--num_migrants", type=int, default=1)
-    parser.add_argument("-pln", "--pollination", action="store_true")
-    parser.add_argument(
-        "-t", "--top_n", type=int, default=1
-    )  # Print top-n best individuals on each island in summary.
-    parser.add_argument("-l", "--logging_int", type=int, default=10)  # Logging interval
-    config = parser.parse_args()
+    # Parse command-line arguments.
+    config, _ = parse_arguments(comm)
 
     # Set up separate logger for Propulate optimization.
     set_logger_config(
-        level=logging.INFO,  # logging level
-        log_file=f"{config.checkpoint}/islands.log",  # logging path
+        level=config.logging_level,  # Logging level
+        log_file=f"{config.checkpoint}/{pathlib.Path(__file__).stem}.log",  # Logging path
         log_to_stdout=True,  # Print log on stdout.
         log_rank=False,  # Do not prepend MPI rank to logging messages.
         colors=True,  # Use colors.
@@ -128,5 +70,7 @@ if __name__ == "__main__":
 
     # Run actual optimization.
     islands.evolve(
-        top_n=config.top_n, logging_interval=config.logging_int, debug=config.verbosity
+        top_n=config.top_n,
+        logging_interval=config.logging_interval,
+        debug=config.verbosity,
     )
